The proxy collection is now a dictionary instead of an array.  This is necessary for acceptable performance in real apps where you need to quickly access proxies for given item specifications.
No longer removes all proxies when the supported outline view gets a -reloadData.  This was found to have the unacceptable side effect of causing the whole outline to collapse and then reappear (with code to do that) whenever an item was edited or deleted.  At this time, there is no mechanism to remove proxies when their represented model item is deleted.  Whether such a mechanism is needed or not is debatable; may depend on the app.

Added more real-life gotchas to the README.md.